Implement frame-by-frame forward + bug fix in video file.
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@160 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/VideoFile.h
@@ -387,7 +387,7 @@ public:
      *
      * @return Aspect ratio of the video.
      */
-    float getStreamAspectRatio() const;
+    double getStreamAspectRatio() const;
     /**
      * Get the frame rate of the movie file opened.
      *
@@ -396,7 +396,7 @@ public:
      *
      * @return frame rate in Hertz (frames per second), 0 if no video is opened.
      */
-    float getFrameRate() const;
+    double getFrameRate() const;
     /**
      * Get the duration of the VideoFile in seconds.
      *
